refactor: Extract packet views to packets/ submodules (Issue #3 Phase 4)
Created packets/ subdirectory with 5 specialized modules:
- decoded.rs (92 lines) - Decoded packet display router
- list.rs (470 lines) - Packet list with filtering and annotations
- statistics.rs (246 lines) - Statistics charts and analysis
- timeline.rs (262 lines) - Timeline visualization
- comparison.rs (302 lines) - Side-by-side packet comparison

Result: core.rs reduced from 2,941 → 1,626 lines (1,315 lines extracted)
Cumulative: 5,020 → 1,626 lines (3,394 lines extracted, 67.6% reduction)

All modules use pub(super) visibility for internal packet module API.
Module coordination through packets/mod.rs with sibling imports in decoded.rs.
