+++ apps/cli/src/tui/ui/core.rs
@@ -10,6 +10,11 @@ use super::overlays::{
     render_confirmation, render_dialog, render_help_overlay, render_notification,
     render_theme_selector,
 };
+use super::packets::comparison::render_packet_comparison;
+use super::packets::decoded::render_decoded_packets;
+use super::packets::list::render_packet_list;
+use super::packets::statistics::render_packet_statistics;
+use super::packets::timeline::render_packet_timeline;
 use super::session::{
     render_session_list, render_session_loading, render_session_manager, render_session_save,
 };
@@ -1622,1320 +1627,5 @@ fn render_error_message(f: &mut Frame, area: Rect, tool_name: &str, error_msg: &
         );
 
     f.render_widget(paragraph, area);
-}
-
-/// Categorize error and provide helpful suggestions
-/// Render decoded packet list for bt_decode
-fn render_decoded_packets(
-    f: &mut Frame,
-    area: Rect,
-    output: &serde_json::Value,
-    packet_list_state: Option<&crate::tui::app::PacketListState>,
-) {
-    use ratatui::{
-        layout::{Constraint, Direction, Layout},
-        style::{Color, Modifier, Style},
-        text::{Line, Span},
-        widgets::{Block, Borders, Paragraph, Wrap},
-    };
-
-    let packets = match output.get("decoded_packets").and_then(|p| p.as_array()) {
-        Some(p) => p,
-        None => {
-            let text = vec![
-                Line::from(""),
-                Line::from("No packets to display"),
-                Line::from(""),
-                Line::from(Span::styled(
-                    "The capture may be empty or the decode limit was 0.",
-                    Style::default().fg(Color::Gray),
-                )),
-            ];
-            let paragraph = Paragraph::new(text)
-                .block(
-                    Block::default()
-                        .borders(Borders::ALL)
-                        .title("Decoded Packets"),
-                )
-                .alignment(ratatui::layout::Alignment::Center);
-            f.render_widget(paragraph, area);
-            return;
-        }
-    };
-
-    let packet_count = packets.len();
-    if packet_count == 0 {
-        let text = vec![
-            Line::from(""),
-            Line::from("No packets to display"),
-            Line::from(""),
-            Line::from(Span::styled(
-                "The capture may be empty or the decode limit was 0.",
-                Style::default().fg(Color::Gray),
-            )),
-        ];
-        let paragraph = Paragraph::new(text)
-            .block(
-                Block::default()
-                    .borders(Borders::ALL)
-                    .title("Decoded Packets"),
-            )
-            .alignment(ratatui::layout::Alignment::Center);
-        f.render_widget(paragraph, area);
-        return;
-    }
-
-    let default_state = crate::tui::app::PacketListState::new();
-    let state = packet_list_state.unwrap_or(&default_state);
-
-    // Dispatch to appropriate view based on view_mode
-    match state.view_mode {
-        crate::tui::app::PacketViewMode::List => {
-            render_packet_list(f, area, packets, state);
-        }
-        crate::tui::app::PacketViewMode::Statistics => {
-            render_packet_statistics(f, area, packets, state);
-        }
-        crate::tui::app::PacketViewMode::Timeline => {
-            render_packet_timeline(f, area, packets, state);
-        }
-        crate::tui::app::PacketViewMode::Comparison => {
-            render_packet_comparison(f, area, packets, state);
-        }
-    }
-}
-
-/// Render packet list view (original table view)
-fn render_packet_list(
-    f: &mut Frame,
-    area: Rect,
-    packets: &[serde_json::Value],
-    state: &crate::tui::app::PacketListState,
-) {
-    use ratatui::{
-        style::{Color, Modifier, Style},
-        text::{Line, Span},
-        widgets::{Block, Borders, Paragraph, Wrap},
-    };
-
-    let packet_count = packets.len();
-
-    // Apply all active filters
-    let filtered_packets: Vec<(usize, &serde_json::Value)> = packets
-        .iter()
-        .enumerate()
-        .filter(|(_, pkt)| {
-            // Follow stream filter (legacy, keeping for compatibility)
-            if let Some(ref mac) = state.follow_mac {
-                if let Some(packet_mac) = pkt.get("mac_address").and_then(|m| m.as_str()) {
-                    if !packet_mac.contains(mac) {
-                        return false;
-                    }
-                } else {
-                    return false;
-                }
-            }
-
-            // Apply filter rules from state.filters
-            state.filters.matches(pkt)
-        })
-        .collect();
-
-    let displayed_count = filtered_packets.len();
-    if displayed_count == 0 {
-        let text = vec![
-            Line::from(""),
-            Line::from("No packets match the current filter"),
-            Line::from(""),
-            Line::from(Span::styled(
-                "Press '/' to modify filters",
-                Style::default().fg(Color::Gray),
-            )),
-        ];
-        let paragraph = Paragraph::new(text)
-            .block(
-                Block::default()
-                    .borders(Borders::ALL)
-                    .title("Decoded Packets (Filtered)"),
-            )
-            .alignment(ratatui::layout::Alignment::Center);
-        f.render_widget(paragraph, area);
-        return;
-    }
-
-    // Calculate visible area
-    let visible_height = area.height.saturating_sub(6) as usize; // Account for borders, header, and footer
-    let start_idx = state.scroll_offset.min(displayed_count.saturating_sub(1));
-    let end_idx = (start_idx + visible_height).min(displayed_count);
-
-    let mut lines = vec![];
-
-    // Header line with indicators
-    lines.push(Line::from(vec![
-        Span::styled("  ", Style::default()), // Space for bookmark/mark indicators
-        Span::styled(
-            " # ",
-            Style::default()
-                .fg(Color::Yellow)
-                .add_modifier(Modifier::BOLD),
-        ),
-        Span::styled(
-            "Time          ",
-            Style::default()
-                .fg(Color::Yellow)
-                .add_modifier(Modifier::BOLD),
-        ),
-        Span::styled(
-            "Ch ",
-            Style::default()
-                .fg(Color::Yellow)
-                .add_modifier(Modifier::BOLD),
-        ),
-        Span::styled(
-            "RSSI ",
-            Style::default()
-                .fg(Color::Yellow)
-                .add_modifier(Modifier::BOLD),
-        ),
-        Span::styled(
-            "Type          ",
-            Style::default()
-                .fg(Color::Yellow)
-                .add_modifier(Modifier::BOLD),
-        ),
-        Span::styled(
-            "MAC Address             ",
-            Style::default()
-                .fg(Color::Yellow)
-                .add_modifier(Modifier::BOLD),
-        ),
-        Span::styled(
-            "Proto ",
-            Style::default()
-                .fg(Color::Yellow)
-                .add_modifier(Modifier::BOLD),
-        ),
-        Span::styled(
-            "Summary",
-            Style::default()
-                .fg(Color::Yellow)
-                .add_modifier(Modifier::BOLD),
-        ),
-    ]));
-
-    lines.push(Line::from(Span::styled(
-        "━".repeat(area.width as usize - 2),
-        Style::default().fg(Color::DarkGray),
-    )));
-
-    // Render visible packets
-    for display_idx in start_idx..end_idx {
-        if let Some((original_idx, packet)) = filtered_packets.get(display_idx) {
-            let is_selected = display_idx == state.selected_index;
-            let is_expanded = state.is_expanded(*original_idx);
-            let is_bookmarked = state.is_bookmarked(*original_idx);
-            let is_marked = state.is_marked_for_comparison(*original_idx);
-
-            // Extract packet fields
-            let frame_num = packet
-                .get("frame_number")
-                .and_then(|v| v.as_str())
-                .unwrap_or("?");
-            let timestamp = packet
-                .get("timestamp")
-                .and_then(|v| v.as_str())
-                .unwrap_or("Unknown");
-            let time_short = timestamp.split(',').last().unwrap_or(timestamp).trim();
-            let time_display = if time_short.len() > 12 {
-                &time_short[time_short.len() - 12..]
-            } else {
-                time_short
-            };
-
-            let channel = packet
-                .get("channel")
-                .and_then(|v| v.as_str())
-                .unwrap_or("?");
-            let rssi = packet.get("rssi").and_then(|v| v.as_str()).unwrap_or("?");
-            let packet_type = packet
-                .get("packet_type")
-                .and_then(|v| v.as_str())
-                .unwrap_or("?");
-            let mac_address = packet
-                .get("mac_address")
-                .and_then(|v| v.as_str())
-                .unwrap_or("N/A");
-            let protocol = packet
-                .get("protocol")
-                .and_then(|v| v.as_str())
-                .unwrap_or("?");
-            let summary = packet.get("summary").and_then(|v| v.as_str()).unwrap_or("");
-
-            // Color based on packet type
-            let type_color = match packet_type {
-                "ADV_IND" | "ADV_NONCONN_IND" | "ADV_SCAN_IND" => Color::Green,
-                "SCAN_REQ" | "SCAN_RSP" => Color::Cyan,
-                "CONNECT_REQ" => Color::Yellow,
-                "DATA" => Color::Blue,
-                _ => Color::White,
-            };
-
-            // Indicators: bookmark, comparison mark, annotation, expand/collapse
-            let bookmark_indicator = if is_bookmarked { "★" } else { " " };
-            let mark_indicator = if is_marked { "●" } else { " " };
-            let has_annotation = state.has_annotation(*original_idx);
-            let annotation_indicator = if has_annotation { "📝" } else { " " };
-            let expand_indicator = if is_selected {
-                if is_expanded {
-                    "▼"
-                } else {
-                    "▶"
-                }
-            } else {
-                " "
-            };
-
-            let bg_color = if is_selected {
-                Color::DarkGray
-            } else {
-                Color::Reset
-            };
-
-            // Main packet row
-            lines.push(Line::from(vec![
-                Span::styled(
-                    bookmark_indicator,
-                    Style::default().fg(Color::Yellow).bg(bg_color),
-                ),
-                Span::styled(
-                    mark_indicator,
-                    Style::default().fg(Color::Magenta).bg(bg_color),
-                ),
-                Span::styled(
-                    expand_indicator,
-                    Style::default().fg(Color::Cyan).bg(bg_color),
-                ),
-                Span::styled(
-                    format!("{:3} ", frame_num),
-                    Style::default().fg(Color::Gray).bg(bg_color),
-                ),
-                Span::styled(
-                    format!("{:12} ", time_display),
-                    Style::default().fg(Color::White).bg(bg_color),
-                ),
-                Span::styled(
-                    format!("{:2} ", channel),
-                    Style::default().fg(Color::Magenta).bg(bg_color),
-                ),
-                Span::styled(
-                    format!("{:4} ", rssi),
-                    Style::default().fg(Color::Red).bg(bg_color),
-                ),
-                Span::styled(
-                    format!("{:13} ", packet_type),
-                    Style::default().fg(type_color).bg(bg_color),
-                ),
-                Span::styled(
-                    format!("{:23} ", mac_address),
-                    Style::default().fg(Color::Cyan).bg(bg_color),
-                ),
-                Span::styled(
-                    format!("{:5} ", protocol),
-                    Style::default().fg(Color::Blue).bg(bg_color),
-                ),
-                Span::styled(summary, Style::default().fg(Color::White).bg(bg_color)),
-            ]));
-
-            // Expanded view
-            if is_expanded {
-                let access_addr = packet
-                    .get("access_addr")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("Unknown");
-
-                lines.push(Line::from(vec![
-                    Span::raw("  │ "),
-                    Span::styled("Access Address: ", Style::default().fg(Color::Gray)),
-                    Span::styled(access_addr, Style::default().fg(Color::White)),
-                ]));
-
-                lines.push(Line::from(vec![
-                    Span::raw("  │ "),
-                    Span::styled("Full timestamp: ", Style::default().fg(Color::Gray)),
-                    Span::styled(timestamp, Style::default().fg(Color::White)),
-                ]));
-
-                // Show protocol layers if available
-                if let Some(full_packet) = packet.get("full_packet") {
-                    if let Some(layers) = full_packet.get("_source").and_then(|s| s.get("layers")) {
-                        let layer_names: Vec<String> = if let Some(obj) = layers.as_object() {
-                            obj.keys().map(|k| k.to_string()).collect()
-                        } else {
-                            vec![]
-                        };
-
-                        if !layer_names.is_empty() {
-                            lines.push(Line::from(vec![
-                                Span::raw("  │ "),
-                                Span::styled("Layers: ", Style::default().fg(Color::Gray)),
-                                Span::styled(
-                                    layer_names.join(" → "),
-                                    Style::default().fg(Color::Cyan),
-                                ),
-                            ]));
-                        }
-                    }
-                }
-
-                // Show annotation if present
-                if let Some(note) = state.get_annotation(*original_idx) {
-                    lines.push(Line::from(vec![
-                        Span::raw("  │ "),
-                        Span::styled(
-                            "Note: ",
-                            Style::default()
-                                .fg(Color::Yellow)
-                                .add_modifier(Modifier::BOLD),
-                        ),
-                        Span::styled(
-                            note.clone(),
-                            Style::default()
-                                .fg(Color::White)
-                                .add_modifier(Modifier::ITALIC),
-                        ),
-                    ]));
-                }
-
-                lines.push(Line::from(vec![
-                    Span::raw("  │ "),
-                    Span::styled(
-                        "[Enter: collapse | n: add note | Del: remove note]",
-                        Style::default()
-                            .fg(Color::DarkGray)
-                            .add_modifier(Modifier::ITALIC),
-                    ),
-                ]));
-
-                lines.push(Line::from("  │"));
-            }
-        }
-    }
-
-    // Footer with navigation hints and stats
-    lines.push(Line::from(""));
-
-    // Navigation line
-    lines.push(Line::from(vec![
-        Span::styled("Nav: ", Style::default().fg(Color::Yellow)),
-        Span::styled("↑↓", Style::default().fg(Color::Cyan)),
-        Span::raw(" scroll  "),
-        Span::styled("Enter", Style::default().fg(Color::Cyan)),
-        Span::raw(" expand  "),
-        Span::styled("b", Style::default().fg(Color::Cyan)),
-        Span::raw(" bookmark  "),
-        Span::styled("m", Style::default().fg(Color::Cyan)),
-        Span::raw(" mark  "),
-        Span::styled("f", Style::default().fg(Color::Cyan)),
-        Span::raw(" follow  "),
-    ]));
-
-    // View mode line
-    lines.push(Line::from(vec![
-        Span::styled("Views: ", Style::default().fg(Color::Yellow)),
-        Span::styled("l", Style::default().fg(Color::Cyan)),
-        Span::raw(" list  "),
-        Span::styled("s", Style::default().fg(Color::Cyan)),
-        Span::raw(" statistics  "),
-        Span::styled("t", Style::default().fg(Color::Cyan)),
-        Span::raw(" timeline  "),
-        Span::styled("c", Style::default().fg(Color::Cyan)),
-        Span::raw(" compare  "),
-        Span::styled("n", Style::default().fg(Color::Cyan)),
-        Span::raw(" note  "),
-        Span::styled("/", Style::default().fg(Color::Cyan)),
-        Span::raw(" filter  "),
-        Span::styled("e", Style::default().fg(Color::Cyan)),
-        Span::raw(" export  "),
-        Span::raw(" │ "),
-        Span::styled(
-            format!(
-                "Showing {}-{} of {}",
-                start_idx + 1,
-                end_idx,
-                displayed_count
-            ),
-            Style::default().fg(Color::Gray),
-        ),
-        if displayed_count < packet_count {
-            Span::styled(
-                format!(" (filtered from {})", packet_count),
-                Style::default().fg(Color::Yellow),
-            )
-        } else {
-            Span::raw("")
-        },
-    ]));
-
-    // Follow stream indicator
-    if let Some(ref mac) = state.follow_mac {
-        lines.push(Line::from(vec![
-            Span::styled("Following: ", Style::default().fg(Color::Yellow)),
-            Span::styled(
-                mac.clone(),
-                Style::default()
-                    .fg(Color::Green)
-                    .add_modifier(Modifier::BOLD),
-            ),
-            Span::raw("  "),
-            Span::styled(
-                "(press 'f' to clear)",
-                Style::default()
-                    .fg(Color::Gray)
-                    .add_modifier(Modifier::ITALIC),
-            ),
-        ]));
-    }
-
-    // Bookmark indicator
-    let bookmark_count = state.bookmarks.len();
-    if bookmark_count > 0 {
-        lines.push(Line::from(vec![
-            Span::styled("★ ", Style::default().fg(Color::Yellow)),
-            Span::styled(
-                format!(
-                    "{} bookmarked packet{}",
-                    bookmark_count,
-                    if bookmark_count == 1 { "" } else { "s" }
-                ),
-                Style::default().fg(Color::White),
-            ),
-        ]));
-    }
-
-    // Active filters indicator
-    if state.filters.is_active() {
-        let mut filter_parts = vec![];
-
-        if !state.filters.packet_types.is_empty() {
-            filter_parts.push(format!("Types: {}", state.filters.packet_types.join(", ")));
-        }
-        if let Some(ref mac) = state.filters.mac_address {
-            filter_parts.push(format!("MAC: {}", mac));
-        }
-        if state.filters.rssi_min.is_some() || state.filters.rssi_max.is_some() {
-            let min = state
-                .filters
-                .rssi_min
-                .map(|v| v.to_string())
-                .unwrap_or_else(|| "?".to_string());
-            let max = state
-                .filters
-                .rssi_max
-                .map(|v| v.to_string())
-                .unwrap_or_else(|| "?".to_string());
-            filter_parts.push(format!("RSSI: {} to {} dBm", min, max));
-        }
-
-        lines.push(Line::from(vec![
-            Span::styled(
-                "🔍 Active Filters: ",
-                Style::default()
-                    .fg(Color::Cyan)
-                    .add_modifier(Modifier::BOLD),
-            ),
-            Span::styled(filter_parts.join(" | "), Style::default().fg(Color::White)),
-            Span::raw("  "),
-            Span::styled(
-                "(press '/' to modify)",
-                Style::default()
-                    .fg(Color::Gray)
-                    .add_modifier(Modifier::ITALIC),
-            ),
-        ]));
-    }
-
-    let title = format!(" Decoded Packets ({} total) ", packet_count);
-    let paragraph = Paragraph::new(lines)
-        .block(Block::default().borders(Borders::ALL).title(title))
-        .wrap(Wrap { trim: false });
-
-    f.render_widget(paragraph, area);
-}
-/// Render packet statistics view
-fn render_packet_statistics(
-    f: &mut Frame,
-    area: Rect,
-    packets: &[serde_json::Value],
-    state: &crate::tui::app::PacketListState,
-) {
-    use ratatui::{
-        layout::{Constraint, Direction, Layout},
-        style::{Color, Modifier, Style},
-        text::{Line, Span},
-        widgets::{Block, Borders, Paragraph, Wrap},
-    };
-    use std::collections::HashMap;
-
-    let packet_count = packets.len();
-
-    // Split area into sections
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Length(12), // Packet type distribution
-            Constraint::Length(8),  // Channel and RSSI stats
-            Constraint::Min(8),     // MAC addresses
-        ])
-        .split(area);
-
-    // 1. Packet Type Distribution
-    let mut type_counts: HashMap<String, usize> = HashMap::new();
-    let mut total_rssi: i32 = 0;
-    let mut rssi_count = 0;
-    let mut channel_counts: HashMap<String, usize> = HashMap::new();
-    let mut mac_addresses: std::collections::HashSet<String> = std::collections::HashSet::new();
-    let mut min_rssi = 0i8;
-    let mut max_rssi = -128i8;
-
-    for packet in packets {
-        // Packet types
-        if let Some(ptype) = packet.get("packet_type").and_then(|v| v.as_str()) {
-            *type_counts.entry(ptype.to_string()).or_insert(0) += 1;
-        }
-
-        // RSSI stats
-        if let Some(rssi_str) = packet.get("rssi").and_then(|v| v.as_str()) {
-            if let Ok(rssi) = rssi_str.parse::<i8>() {
-                total_rssi += rssi as i32;
-                rssi_count += 1;
-                min_rssi = min_rssi.min(rssi);
-                max_rssi = max_rssi.max(rssi);
-            }
-        }
-
-        // Channel distribution
-        if let Some(ch) = packet.get("channel").and_then(|v| v.as_str()) {
-            *channel_counts.entry(ch.to_string()).or_insert(0) += 1;
-        }
-
-        // MAC addresses
-        if let Some(mac) = packet.get("mac_address").and_then(|v| v.as_str()) {
-            // Split combined MACs (e.g., "AA ← BB")
-            for part in mac.split(&['←', '→'][..]) {
-                let cleaned = part.trim();
-                if !cleaned.is_empty() && cleaned != "N/A" {
-                    mac_addresses.insert(cleaned.to_string());
-                }
-            }
-        }
-    }
-
-    let avg_rssi = if rssi_count > 0 {
-        total_rssi as f32 / rssi_count as f32
-    } else {
-        0.0
-    };
-
-    // Render packet type distribution
-    let mut type_lines = vec![
-        Line::from(Span::styled(
-            "Packet Type Distribution",
-            Style::default()
-                .fg(Color::Yellow)
-                .add_modifier(Modifier::BOLD),
-        )),
-        Line::from(""),
-    ];
-
-    let mut sorted_types: Vec<_> = type_counts.iter().collect();
-    sorted_types.sort_by(|a, b| b.1.cmp(a.1));
-
-    for (ptype, count) in sorted_types.iter().take(8) {
-        let percentage = (**count as f32 / packet_count as f32) * 100.0;
-        let bar_width = (percentage / 3.0) as usize; // Scale for display
-        let bar = "█".repeat(bar_width.max(1));
-
-        let color = match ptype.as_str() {
-            "ADV_IND" | "ADV_NONCONN_IND" | "ADV_SCAN_IND" => Color::Green,
-            "SCAN_REQ" | "SCAN_RSP" => Color::Cyan,
-            "CONNECT_REQ" => Color::Yellow,
-            "DATA" => Color::Blue,
-            _ => Color::White,
-        };
-
-        type_lines.push(Line::from(vec![
-            Span::styled(format!("{:15}", ptype), Style::default().fg(Color::White)),
-            Span::styled(bar, Style::default().fg(color)),
-            Span::styled(
-                format!(" {} ({:.1}%)", count, percentage),
-                Style::default().fg(Color::Gray),
-            ),
-        ]));
-    }
-
-    let type_block = Paragraph::new(type_lines)
-        .block(
-            Block::default()
-                .borders(Borders::ALL)
-                .title(" Packet Types "),
-        )
-        .wrap(Wrap { trim: false });
-    f.render_widget(type_block, chunks[0]);
-
-    // Render channel and RSSI stats
-    let mut stats_lines = vec![
-        Line::from(vec![
-            Span::styled("Total Packets: ", Style::default().fg(Color::Yellow)),
-            Span::styled(
-                packet_count.to_string(),
-                Style::default()
-                    .fg(Color::White)
-                    .add_modifier(Modifier::BOLD),
-            ),
-        ]),
-        Line::from(vec![
-            Span::styled("Unique MACs:   ", Style::default().fg(Color::Yellow)),
-            Span::styled(
-                mac_addresses.len().to_string(),
-                Style::default()
-                    .fg(Color::White)
-                    .add_modifier(Modifier::BOLD),
-            ),
-        ]),
-        Line::from(""),
-        Line::from(vec![
-            Span::styled("RSSI Stats:    ", Style::default().fg(Color::Yellow)),
-            Span::styled(
-                format!("Avg: {:.1} dBm", avg_rssi),
-                Style::default().fg(Color::White),
-            ),
-            Span::raw("  "),
-            Span::styled(
-                format!("Min: {} dBm", min_rssi),
-                Style::default().fg(Color::Red),
-            ),
-            Span::raw("  "),
-            Span::styled(
-                format!("Max: {} dBm", max_rssi),
-                Style::default().fg(Color::Green),
-            ),
-        ]),
-        Line::from(""),
-    ];
-
-    // Channel distribution
-    stats_lines.push(Line::from(Span::styled(
-        "Channel Distribution:",
-        Style::default().fg(Color::Yellow),
-    )));
-
-    let mut sorted_channels: Vec<_> = channel_counts.iter().collect();
-    sorted_channels.sort_by_key(|(ch, _)| ch.parse::<u8>().unwrap_or(255));
-
-    for (ch, count) in sorted_channels {
-        let percentage = (*count as f32 / packet_count as f32) * 100.0;
-        stats_lines.push(Line::from(vec![
-            Span::styled(
-                format!("  Ch {:2}:", ch),
-                Style::default().fg(Color::Magenta),
-            ),
-            Span::styled(
-                format!(" {} packets ({:.1}%)", count, percentage),
-                Style::default().fg(Color::White),
-            ),
-        ]));
-    }
-
-    let stats_block = Paragraph::new(stats_lines)
-        .block(Block::default().borders(Borders::ALL).title(" Statistics "))
-        .wrap(Wrap { trim: false });
-    f.render_widget(stats_block, chunks[1]);
-
-    // Render unique MAC addresses
-    let mut mac_lines = vec![
-        Line::from(Span::styled(
-            format!("Unique MAC Addresses ({})", mac_addresses.len()),
-            Style::default()
-                .fg(Color::Yellow)
-                .add_modifier(Modifier::BOLD),
-        )),
-        Line::from(""),
-    ];
-
-    let mut sorted_macs: Vec<_> = mac_addresses.iter().collect();
-    sorted_macs.sort();
-
-    for (i, mac) in sorted_macs.iter().take(15).enumerate() {
-        mac_lines.push(Line::from(vec![
-            Span::styled(format!("{:2}. ", i + 1), Style::default().fg(Color::Gray)),
-            Span::styled(mac.to_string(), Style::default().fg(Color::Cyan)),
-        ]));
-    }
-
-    if sorted_macs.len() > 15 {
-        mac_lines.push(Line::from(Span::styled(
-            format!("... and {} more", sorted_macs.len() - 15),
-            Style::default()
-                .fg(Color::Gray)
-                .add_modifier(Modifier::ITALIC),
-        )));
-    }
-
-    mac_lines.push(Line::from(""));
-    mac_lines.push(Line::from(vec![
-        Span::styled("Press ", Style::default().fg(Color::Gray)),
-        Span::styled("l", Style::default().fg(Color::Cyan)),
-        Span::styled(
-            " to return to packet list view",
-            Style::default().fg(Color::Gray),
-        ),
-    ]));
-
-    let mac_block = Paragraph::new(mac_lines)
-        .block(Block::default().borders(Borders::ALL).title(" Devices "))
-        .wrap(Wrap { trim: false });
-    f.render_widget(mac_block, chunks[2]);
-}
-/// Render packet timeline view
-fn render_packet_timeline(
-    f: &mut Frame,
-    area: Rect,
-    packets: &[serde_json::Value],
-    state: &crate::tui::app::PacketListState,
-) {
-    use ratatui::{
-        layout::{Constraint, Direction, Layout},
-        style::{Color, Modifier, Style},
-        text::{Line, Span},
-        widgets::{Block, Borders, Paragraph, Wrap},
-    };
-
-    let packet_count = packets.len();
-    if packet_count == 0 {
-        return;
-    }
-
-    // Split area
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Min(10),   // Timeline
-            Constraint::Length(8), // Legend and stats
-        ])
-        .split(area);
-
-    // Parse timestamps and determine time range
-    let mut timestamps: Vec<(usize, f64)> = Vec::new();
-    for (idx, packet) in packets.iter().enumerate() {
-        if let Some(ts_str) = packet.get("timestamp").and_then(|v| v.as_str()) {
-            // Parse timestamp - just use index if parsing fails
-            // Format: "Mar  4, 2026 17:07:00.999844945 EST"
-            timestamps.push((idx, idx as f64));
-        }
-    }
-
-    if timestamps.is_empty() {
-        return;
-    }
-
-    let min_time = 0.0;
-    let max_time = packet_count as f64;
-    let time_range = max_time - min_time;
-
-    // Calculate timeline width
-    let timeline_width = area.width.saturating_sub(6) as usize;
-    let timeline_height = chunks[0].height.saturating_sub(4) as usize;
-
-    // Create timeline buckets
-    let bucket_count = timeline_width;
-    let mut buckets: Vec<Vec<(usize, &serde_json::Value)>> = vec![vec![]; bucket_count];
-
-    for (idx, _) in &timestamps {
-        if let Some(packet) = packets.get(*idx) {
-            let normalized_time = (*idx as f64 - min_time) / time_range;
-            let bucket_idx = (normalized_time * (bucket_count - 1) as f64) as usize;
-            if bucket_idx < bucket_count {
-                buckets[bucket_idx].push((*idx, packet));
-            }
-        }
-    }
-
-    // Render timeline
-    let mut timeline_lines = vec![];
-
-    // Title
-    timeline_lines.push(Line::from(vec![
-        Span::styled(
-            "Packet Activity Over Time",
-            Style::default()
-                .fg(Color::Yellow)
-                .add_modifier(Modifier::BOLD),
-        ),
-        Span::raw("  "),
-        Span::styled(
-            format!("({} packets)", packet_count),
-            Style::default().fg(Color::Gray),
-        ),
-    ]));
-    timeline_lines.push(Line::from(""));
-
-    // Draw timeline rows (different packet types stacked)
-    let packet_types = vec!["ADV_IND", "SCAN_REQ", "SCAN_RSP", "CONNECT_REQ", "DATA"];
-    let type_colors = vec![
-        Color::Green,
-        Color::Cyan,
-        Color::Blue,
-        Color::Yellow,
-        Color::Magenta,
-    ];
-
-    for (type_name, color) in packet_types.iter().zip(type_colors.iter()) {
-        let mut line_spans = vec![Span::styled(
-            format!("{:12} ", type_name),
-            Style::default().fg(*color),
-        )];
-
-        for bucket in &buckets {
-            let count_of_type = bucket
-                .iter()
-                .filter(|(_, pkt)| {
-                    pkt.get("packet_type")
-                        .and_then(|v| v.as_str())
-                        .map(|t| t == *type_name)
-                        .unwrap_or(false)
-                })
-                .count();
-
-            let symbol = if count_of_type == 0 {
-                "·"
-            } else if count_of_type == 1 {
-                "▁"
-            } else if count_of_type == 2 {
-                "▃"
-            } else if count_of_type <= 4 {
-                "▅"
-            } else {
-                "█"
-            };
-
-            line_spans.push(Span::styled(symbol, Style::default().fg(*color)));
-        }
-
-        timeline_lines.push(Line::from(line_spans));
-    }
-
-    // Add a separator
-    timeline_lines.push(Line::from(""));
-    timeline_lines.push(Line::from(Span::styled(
-        "─".repeat(timeline_width + 15),
-        Style::default().fg(Color::DarkGray),
-    )));
-
-    // Density visualization (all packet types combined)
-    timeline_lines.push(Line::from(""));
-    timeline_lines.push(Line::from(Span::styled(
-        "All Packets  ",
-        Style::default()
-            .fg(Color::White)
-            .add_modifier(Modifier::BOLD),
-    )));
-
-    let mut density_line = vec![Span::raw("             ")];
-    for bucket in &buckets {
-        let total = bucket.len();
-        let symbol = if total == 0 {
-            " "
-        } else if total == 1 {
-            "░"
-        } else if total <= 3 {
-            "▒"
-        } else if total <= 6 {
-            "▓"
-        } else {
-            "█"
-        };
-        density_line.push(Span::styled(symbol, Style::default().fg(Color::White)));
-    }
-    timeline_lines.push(Line::from(density_line));
-
-    // Time axis markers
-    timeline_lines.push(Line::from(""));
-    let mut axis_line = vec![Span::raw("             ")];
-    for i in 0..bucket_count {
-        if i % 10 == 0 {
-            axis_line.push(Span::styled("|", Style::default().fg(Color::Gray)));
-        } else {
-            axis_line.push(Span::raw(" "));
-        }
-    }
-    timeline_lines.push(Line::from(axis_line));
-
-    // Time labels
-    let mut label_line = vec![Span::styled(
-        "Time -->     ",
-        Style::default().fg(Color::Gray),
-    )];
-    for i in 0..bucket_count {
-        if i % 20 == 0 && i > 0 {
-            let packet_num = ((i as f64 / bucket_count as f64) * packet_count as f64) as usize;
-            label_line.push(Span::styled(
-                format!("{}", packet_num),
-                Style::default().fg(Color::Gray),
-            ));
-            // Add spacing
-            for _ in 0..format!("{}", packet_num).len() {
-                if i + 1 < bucket_count {
-                    label_line.push(Span::raw(" "));
-                }
-            }
-        }
-    }
-    timeline_lines.push(Line::from(label_line));
-
-    let timeline_block = Paragraph::new(timeline_lines)
-        .block(
-            Block::default()
-                .borders(Borders::ALL)
-                .title(" Timeline View "),
-        )
-        .wrap(Wrap { trim: false });
-    f.render_widget(timeline_block, chunks[0]);
-
-    // Legend and navigation
-    let legend_lines = vec![
-        Line::from(vec![
-            Span::styled("Legend: ", Style::default().fg(Color::Yellow)),
-            Span::styled("· ", Style::default().fg(Color::DarkGray)),
-            Span::raw("none  "),
-            Span::styled("▁ ", Style::default().fg(Color::White)),
-            Span::raw("1  "),
-            Span::styled("▃ ", Style::default().fg(Color::White)),
-            Span::raw("2  "),
-            Span::styled("▅ ", Style::default().fg(Color::White)),
-            Span::raw("3-4  "),
-            Span::styled("█ ", Style::default().fg(Color::White)),
-            Span::raw("5+  "),
-        ]),
-        Line::from(""),
-        Line::from(vec![
-            Span::styled("Packet Types: ", Style::default().fg(Color::Yellow)),
-            Span::styled("■ ", Style::default().fg(Color::Green)),
-            Span::raw("ADV  "),
-            Span::styled("■ ", Style::default().fg(Color::Cyan)),
-            Span::raw("SCAN_REQ  "),
-            Span::styled("■ ", Style::default().fg(Color::Blue)),
-            Span::raw("SCAN_RSP  "),
-            Span::styled("■ ", Style::default().fg(Color::Yellow)),
-            Span::raw("CONNECT  "),
-            Span::styled("■ ", Style::default().fg(Color::Magenta)),
-            Span::raw("DATA"),
-        ]),
-        Line::from(""),
-        Line::from(vec![
-            Span::styled("Navigation: ", Style::default().fg(Color::Yellow)),
-            Span::styled("l", Style::default().fg(Color::Cyan)),
-            Span::raw(" list view  "),
-            Span::styled("s", Style::default().fg(Color::Cyan)),
-            Span::raw(" statistics  "),
-            Span::styled("t", Style::default().fg(Color::Cyan)),
-            Span::raw(" timeline"),
-        ]),
-    ];
-
-    let legend_block = Paragraph::new(legend_lines)
-        .block(Block::default().borders(Borders::ALL).title(" Legend "))
-        .wrap(Wrap { trim: false });
-    f.render_widget(legend_block, chunks[1]);
-}
-/// Render side-by-side packet comparison view
-fn render_packet_comparison(
-    f: &mut Frame,
-    area: Rect,
-    packets: &[serde_json::Value],
-    state: &crate::tui::app::PacketListState,
-) {
-    use ratatui::{
-        layout::{Constraint, Direction, Layout},
-        style::{Color, Modifier, Style},
-        text::{Line, Span},
-        widgets::{Block, Borders, Paragraph, Wrap},
-    };
-
-    // Check if we have exactly 2 packets marked for comparison
-    if state.comparison_marks.len() != 2 {
-        let text = vec![
-            Line::from(""),
-            Line::from(Span::styled(
-                "Please mark exactly 2 packets for comparison",
-                Style::default()
-                    .fg(Color::Yellow)
-                    .add_modifier(Modifier::BOLD),
-            )),
-            Line::from(""),
-            Line::from(vec![
-                Span::styled("1. ", Style::default().fg(Color::Gray)),
-                Span::raw("Select a packet and press "),
-                Span::styled(
-                    "m",
-                    Style::default()
-                        .fg(Color::Cyan)
-                        .add_modifier(Modifier::BOLD),
-                ),
-                Span::raw(" to mark it"),
-            ]),
-            Line::from(vec![
-                Span::styled("2. ", Style::default().fg(Color::Gray)),
-                Span::raw("Select another packet and press "),
-                Span::styled(
-                    "m",
-                    Style::default()
-                        .fg(Color::Cyan)
-                        .add_modifier(Modifier::BOLD),
-                ),
-                Span::raw(" again"),
-            ]),
-            Line::from(vec![
-                Span::styled("3. ", Style::default().fg(Color::Gray)),
-                Span::raw("Press "),
-                Span::styled(
-                    "c",
-                    Style::default()
-                        .fg(Color::Cyan)
-                        .add_modifier(Modifier::BOLD),
-                ),
-                Span::raw(" to open comparison view"),
-            ]),
-            Line::from(""),
-            Line::from(vec![
-                Span::styled("Currently marked: ", Style::default().fg(Color::Yellow)),
-                Span::styled(
-                    format!("{} packet(s)", state.comparison_marks.len()),
-                    Style::default().fg(Color::White),
-                ),
-            ]),
-            Line::from(""),
-            Line::from(vec![
-                Span::styled("Press ", Style::default().fg(Color::Gray)),
-                Span::styled("l", Style::default().fg(Color::Cyan)),
-                Span::styled(" to return to list view", Style::default().fg(Color::Gray)),
-            ]),
-        ];
-
-        let paragraph = Paragraph::new(text)
-            .block(
-                Block::default()
-                    .borders(Borders::ALL)
-                    .title(" Packet Comparison "),
-            )
-            .alignment(ratatui::layout::Alignment::Center);
-        f.render_widget(paragraph, area);
-        return;
-    }
-
-    // Get the two marked packets
-    let idx1 = state.comparison_marks[0];
-    let idx2 = state.comparison_marks[1];
-
-    let packet1 = packets.get(idx1);
-    let packet2 = packets.get(idx2);
-
-    if packet1.is_none() || packet2.is_none() {
-        let text = vec![
-            Line::from(""),
-            Line::from("Error: Could not load marked packets"),
-            Line::from(""),
-        ];
-        let paragraph = Paragraph::new(text)
-            .block(Block::default().borders(Borders::ALL).title(" Error "))
-            .alignment(ratatui::layout::Alignment::Center);
-        f.render_widget(paragraph, area);
-        return;
-    }
-
-    let packet1 = packet1.unwrap();
-    let packet2 = packet2.unwrap();
-
-    // Split screen into two columns
-    let chunks = Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([Constraint::Percentage(50), Constraint::Percentage(50)])
-        .split(area);
-
-    // Helper function to extract field with default
-    let get_field = |pkt: &serde_json::Value, field: &str, default: &str| {
-        pkt.get(field)
-            .and_then(|v| v.as_str())
-            .unwrap_or(default)
-            .to_string()
-    };
-
-    // Helper function to render one packet
-    let render_packet_details = |pkt: &serde_json::Value, idx: usize, area: Rect, f: &mut Frame| {
-        let frame_num = get_field(pkt, "frame_number", "?");
-        let timestamp = get_field(pkt, "timestamp", "Unknown");
-        let channel = get_field(pkt, "channel", "?");
-        let rssi = get_field(pkt, "rssi", "?");
-        let packet_type = get_field(pkt, "packet_type", "?");
-        let mac_address = get_field(pkt, "mac_address", "N/A");
-        let protocol = get_field(pkt, "protocol", "?");
-        let summary = get_field(pkt, "summary", "");
-        let access_addr = get_field(pkt, "access_addr", "Unknown");
-
-        // Determine if fields differ
-        let p1_frame = get_field(packet1, "frame_number", "?");
-        let p2_frame = get_field(packet2, "frame_number", "?");
-        let p1_channel = get_field(packet1, "channel", "?");
-        let p2_channel = get_field(packet2, "channel", "?");
-        let p1_rssi = get_field(packet1, "rssi", "?");
-        let p2_rssi = get_field(packet2, "rssi", "?");
-        let p1_type = get_field(packet1, "packet_type", "?");
-        let p2_type = get_field(packet2, "packet_type", "?");
-        let p1_mac = get_field(packet1, "mac_address", "N/A");
-        let p2_mac = get_field(packet2, "mac_address", "N/A");
-        let p1_protocol = get_field(packet1, "protocol", "?");
-        let p2_protocol = get_field(packet2, "protocol", "?");
-        let p1_access = get_field(packet1, "access_addr", "Unknown");
-        let p2_access = get_field(packet2, "access_addr", "Unknown");
-
-        let differ_channel = p1_channel != p2_channel;
-        let differ_rssi = p1_rssi != p2_rssi;
-        let differ_type = p1_type != p2_type;
-        let differ_mac = p1_mac != p2_mac;
-        let differ_protocol = p1_protocol != p2_protocol;
-        let differ_access = p1_access != p2_access;
-
-        let diff_style = Style::default()
-            .fg(Color::Yellow)
-            .add_modifier(Modifier::BOLD);
-        let same_style = Style::default().fg(Color::White);
-
-        let mut lines = vec![
-            Line::from(Span::styled(
-                format!("Packet #{} (Index: {})", frame_num, idx),
-                Style::default()
-                    .fg(Color::Cyan)
-                    .add_modifier(Modifier::BOLD),
-            )),
-            Line::from(""),
-            Line::from(vec![
-                Span::styled("Timestamp:    ", Style::default().fg(Color::Gray)),
-                Span::styled(timestamp, Style::default().fg(Color::White)),
-            ]),
-            Line::from(""),
-            Line::from(vec![
-                Span::styled("Channel:      ", Style::default().fg(Color::Gray)),
-                Span::styled(
-                    channel,
-                    if differ_channel {
-                        diff_style
-                    } else {
-                        same_style
-                    },
-                ),
-            ]),
-            Line::from(vec![
-                Span::styled("RSSI:         ", Style::default().fg(Color::Gray)),
-                Span::styled(
-                    format!("{} dBm", rssi),
-                    if differ_rssi { diff_style } else { same_style },
-                ),
-            ]),
-            Line::from(vec![
-                Span::styled("Type:         ", Style::default().fg(Color::Gray)),
-                Span::styled(
-                    packet_type.clone(),
-                    if differ_type { diff_style } else { same_style },
-                ),
-            ]),
-            Line::from(vec![
-                Span::styled("Protocol:     ", Style::default().fg(Color::Gray)),
-                Span::styled(
-                    protocol,
-                    if differ_protocol {
-                        diff_style
-                    } else {
-                        same_style
-                    },
-                ),
-            ]),
-            Line::from(""),
-            Line::from(vec![Span::styled(
-                "MAC Address:  ",
-                Style::default().fg(Color::Gray),
-            )]),
-            Line::from(vec![
-                Span::raw("  "),
-                Span::styled(
-                    mac_address,
-                    if differ_mac { diff_style } else { same_style },
-                ),
-            ]),
-            Line::from(""),
-            Line::from(vec![
-                Span::styled("Access Addr:  ", Style::default().fg(Color::Gray)),
-                Span::styled(
-                    access_addr,
-                    if differ_access {
-                        diff_style
-                    } else {
-                        same_style
-                    },
-                ),
-            ]),
-            Line::from(""),
-            Line::from(vec![Span::styled(
-                "Summary:      ",
-                Style::default().fg(Color::Gray),
-            )]),
-            Line::from(vec![
-                Span::raw("  "),
-                Span::styled(
-                    summary,
-                    Style::default()
-                        .fg(Color::White)
-                        .add_modifier(Modifier::ITALIC),
-                ),
-            ]),
-        ];
-
-        // Add annotation if present
-        if let Some(note) = state.get_annotation(idx) {
-            lines.push(Line::from(""));
-            lines.push(Line::from(vec![Span::styled(
-                "Note:         ",
-                Style::default()
-                    .fg(Color::Yellow)
-                    .add_modifier(Modifier::BOLD),
-            )]));
-            lines.push(Line::from(vec![
-                Span::raw("  "),
-                Span::styled(
-                    note.clone(),
-                    Style::default()
-                        .fg(Color::White)
-                        .add_modifier(Modifier::ITALIC),
-                ),
-            ]));
-        }
-
-        lines.push(Line::from(""));
-        lines.push(Line::from(""));
-        lines.push(Line::from(Span::styled(
-            "Differences highlighted in yellow",
-            Style::default()
-                .fg(Color::DarkGray)
-                .add_modifier(Modifier::ITALIC),
-        )));
-
-        let title = format!(" Packet {} ", if idx == idx1 { "A" } else { "B" });
-        let paragraph = Paragraph::new(lines)
-            .block(Block::default().borders(Borders::ALL).title(title))
-            .wrap(Wrap { trim: false });
-        f.render_widget(paragraph, area);
-    };
-
-    // Render both packets side by side
-    render_packet_details(packet1, idx1, chunks[0], f);
-    render_packet_details(packet2, idx2, chunks[1], f);
 }
