+++ apps/cli/src/tui/ui/packets/mod.rs
@@ -1,0 +1,7 @@
+//! Packet view rendering modules
+
+pub(super) mod comparison;
+pub(super) mod decoded;
+pub(super) mod list;
+pub(super) mod statistics;
+pub(super) mod timeline;
